+++ thrift/client/src/SENNA_utils.h
@@ -16,8 +16,8 @@ FILE *SENNA_fopen(const char *path, const char *subpath, const char *mode);
 void SENNA_fseek(FILE *stream, long offset, int whence);
 long SENNA_ftell(FILE *stream);
 void SENNA_fread(void *ptr, size_t size, size_t nmemb, FILE *stream);
-void SENNA_fread_tensor_1d(double **ptr, int *n_row, FILE *stream);
-void SENNA_fread_tensor_2d(double **ptr, int *n_row, int *n_column, FILE *stream);
+void SENNA_fread_tensor_1d(float **ptr, int *n_row, FILE *stream);
+void SENNA_fread_tensor_2d(float **ptr, int *n_row, int *n_column, FILE *stream);
 char* SENNA_fgetline(char *str, int size, FILE *stream);
 void SENNA_fclose(FILE *stream);
 
@@ -27,7 +27,7 @@ void* SENNA_realloc(void *ptr, size_t size, size_t nitems);
 void SENNA_free(void *ptr);
 
 /* debug */
-void SENNA_print_tensor_1d(double *tensor, int nrow);
-void SENNA_print_tensor_2d(double *tensor, int nrow, int ncolumn);
+void SENNA_print_tensor_1d(float *tensor, int nrow);
+void SENNA_print_tensor_2d(float *tensor, int nrow, int ncolumn);
 
 #endif
